+++ R/tarpuy_design.R
@@ -8,10 +8,11 @@
 #'   details.
 #' @param type Type of experimental arrange  (default = "crd"). See details.
 #' @param rep  Number of replications in the experiment (default = 3).
+#' @param zigzag Experiment layout in zigzag [logic: F].
 #' @param serie Digits in the plot id (default = 2).
 #' @param seed Replicability of draw results (default = 0) always random. See
 #'   details.
-#' @param barcode Barcode prefix for data collection.
+#' @param fbname Barcode prefix for data collection.
 #'
 #' @details The function allows to include the arguments in the sheet that have
 #'   the information of the design. You should include 2 columns in the sheet:
@@ -40,12 +41,17 @@
 #' library(gsheet)
 #' 
 #' url <- paste0("https://docs.google.com/spreadsheets/d/"
-#'               , "1tDcLZZ9C80uBJK8RRW6_WIyxiL8bIHACStJyf8avuCU/edit#gid=59220990")
+#'               , "1grAv_2po804pPGg9nj1o5nli01IcEGvSevDruq_ssHk/edit#gid=244058650")
 #' # browseURL(url)
 #' 
 #' fb <- gsheet2tbl(url) 
 #' 
-#' tarpuy_design(data = fb) 
+#' dsg <- fb %>% tarpuy_design() 
+#' 
+#' dsg
+#' 
+#' dsg %>% 
+#'   tarpuy_plotdesign()
 #' 
 #' }
 
@@ -53,12 +59,13 @@ tarpuy_design <- function(data
                           , nfactors = 1
                           , type = "crd"
                           , rep = 2
+                          , zigzag = FALSE
                           , serie = 2
-                          , seed = 0
-                          , barcode = NA
-                             ) {
+                          , seed = NULL
+                          , fbname = NA
+                          ) {
 
-  plots <- Row.names <- factors <- where <- NULL
+plots <- Row.names <- factors <- where <- NULL
   
 # design type -------------------------------------------------------------
 # -------------------------------------------------------------------------
@@ -71,6 +78,8 @@ type <- match.arg(type, c(
 
 # factors -----------------------------------------------------------------
 # -------------------------------------------------------------------------
+
+# data <- fb
 
 dt_factors <- data %>%
   dplyr::select(where(~!all(is.na(.)))) %>% 
@@ -108,7 +117,7 @@ opt <- list(nfactors = nfactors
             , rep = rep
             , serie = serie
             , seed = seed
-            , barcode = barcode
+            , fbname = fbname
             ) %>% 
   tibble::enframe(name = "{arguments}", value = "{values}") %>% 
   merge(.
@@ -144,20 +153,21 @@ rep <- if(is.null(arguments$rep) || is.na(arguments$rep) || arguments$rep == "")
 } else {arguments$rep} %>% 
   as.numeric()
 
+zigzag <- if(is.null(arguments$zigzag) || is.na(arguments$zigzag) || arguments$zigzag == "") { zigzag
+} else {arguments$zigzag} %>% as.logical()
+
 serie <- if(is.null(arguments$serie) || is.na(arguments$serie) || arguments$serie == "") { serie
 } else {arguments$serie} %>% 
   as.numeric()
 
-seed <- if(is.null(arguments$seed) || is.na(arguments$seed) || arguments$seed == "") { seed
-} else {arguments$seed} %>% 
-  as.numeric()
-
-barcode <- if(is.null(arguments$barcode) || is.na(arguments$barcode) || arguments$barcode == "") { barcode
-} else {arguments$barcode} %>% 
+seed <- if(is.null(arguments$seed) || is.na(arguments$seed) || arguments$seed == "" || arguments$seed == "0") { NULL
+} else {arguments$seed %>% as.numeric()} 
+  
+fbname <- if(is.null(arguments$fbname) || is.na(arguments$fbname) || arguments$fbname == "") { fbname
+} else {arguments$fbname} %>% 
   stringi::stri_trans_general("Latin-ASCII") %>%
   stringr::str_to_upper() %>% 
   gsub("[[:space:]]", "-", .)
-
 
 # -------------------------------------------------------------------------
 
@@ -181,7 +191,6 @@ factor_levels <- dt_factors %>%
   purrr::map(stats::na.omit) %>% 
   purrr::map(as.vector)
   
-
 # n_factor = 1 ------------------------------------------------------------
 # -------------------------------------------------------------------------
 
@@ -189,35 +198,24 @@ factor_levels <- dt_factors %>%
             
             factors <- factor_levels[[1]]
             
-            if (type == "crd") {
+            if (type == "crd" || type == "rcbd") {
               
-              design <- agricolae::design.crd(
-                trt = factors
-                , r = rep
+              design <- design_repblock(
+                nfactors = 1
+                , factors = factor_levels
+                , type = type
+                , rep = rep
+                , zigzag = zigzag
                 , serie = serie
                 , seed = seed
-                )
+                , fbname = fbname
+              )
 
-              fb_info <- design$book %>%
-                dplyr::rename( {{factor_names}} := "factors")
+             design$fieldbook 
                 
             }
 
-            if (type == "rcbd") {
-              
-              design <- agricolae::design.rcbd(
-                trt = factors
-                , r = rep
-                , serie = serie
-                , seed = seed
-                )
-
-              fb_info <- design$book %>%
-                dplyr::rename( {{factor_names}} := "factors")
-              
-            }
-
-            if (type == "lsd") {
+            if (type == "lsd") { # need update
 
               design <- agricolae::design.lsd(
                 trt = factors
@@ -229,9 +227,14 @@ factor_levels <- dt_factors %>%
               fb_info <- design$book %>%
                 dplyr::rename( {{factor_names}} := "factors")
               
+              fb <- fb_info %>% 
+                dplyr::mutate(exp = fbname, .before = 1) %>% 
+                tidyr::unite(col = "fbname", everything(), remove = F, sep = "_") %>% 
+                dplyr::select(!exp)
+              
             }
 
-            if (type == "lattice") { # fix rename column?
+            if (type == "lattice") {  # need update
 
               if( rep > 3 ) { rep <- 3 }
 
@@ -244,6 +247,11 @@ factor_levels <- dt_factors %>%
 
               fb_info <- design$book %>%
                 dplyr::rename( {{factor_names}} := "trt")
+              
+              fb <- fb_info %>% 
+                dplyr::mutate(exp = fbname, .before = 1) %>% 
+                tidyr::unite(col = "fbname", everything(), remove = F, sep = "_") %>% 
+                dplyr::select(!exp)
 
             }
 
@@ -252,7 +260,7 @@ factor_levels <- dt_factors %>%
 # n_factor >= 2 -----------------------------------------------------------
 # -------------------------------------------------------------------------
 
-        if( nfactors == 2 & startsWith(type, "split") ) {
+    fb <- if( nfactors == 2 & startsWith(type, "split") ) {  # need update
           
           model <- type %>% gsub("split-", "", .)
             
@@ -271,61 +279,33 @@ factor_levels <- dt_factors %>%
             fb_info <- design$book %>%
               dplyr::rename_with(~ {{ factor_names }}, tail(names(.), 2))
             
+            fb <- fb_info %>% 
+              dplyr::mutate(exp = fbname, .before = 1) %>% 
+              tidyr::unite(col = "fbname", everything(), remove = F, sep = "_") %>% 
+              dplyr::select(!exp)
+            
         }
 
 # factorial ---------------------------------------------------------------
 # -------------------------------------------------------------------------
 
-        if ( nfactors >= 2 & ( type == "crd" | type == "rcbd" | type == "lsd" ) ) {
+    fb <- if ( nfactors >= 2 & ( type == "crd" | type == "rcbd" | type == "lsd" ) ) {
           
-          factors <- lengths(factor_levels)
-
-          design <- agricolae::design.ab(
-            trt = factors
-            , r = rep
+          design <- design_repblock(
+            nfactors = nfactors
+            , factors = factor_levels
+            , type = type
+            , rep = rep
+            , zigzag = zigzag
             , serie = serie
-            , design = type
             , seed = seed
+            , fbname = fbname
           )
           
-          
-          fb <- design$book %>%
-            dplyr::rename_with(~ {{ factor_names }}, tail(names(.), nfactors)) 
-          
-          fb_factors <- fb %>% 
-            dplyr::select(!{{factor_names}}) %>% 
-            names()
-          
-          fb_info <- 1:length(factor_names) %>% purrr::map( function(x) {
-            
-            factor <- names(factors)[x]
-            levels <- factor_levels[[factor]] 
-            
-            org_names <- fb %>% 
-              dplyr::select({{factor}}) %>% 
-              unique() %>%  
-              tibble::deframe() %>% 
-              as.vector()
-            
-            names <- structure(as.character(levels), names = as.character(org_names))
-            
-            fb %>% 
-              dplyr::select(fb_factors, {{factor}}) %>% 
-              dplyr::mutate(across({{factor}}, ~dplyr::recode(.x = ., !!!names)))
-            
-          }) %>% 
-            Reduce(function(...) merge(..., all=T), .)
+          design$fieldbook 
           
         }
           
-# include qr --------------------------------------------------------------
-# -------------------------------------------------------------------------
-
-fb <- fb_info %>% 
-  dplyr::mutate(exp = barcode, .before = 1) %>% 
-  tidyr::unite(col = "barcode", everything(), remove = F, sep = "_") %>% 
-  dplyr::select(!exp)
-
 list(fieldbook = fb)
   
 }