+++ NEWS.md
@@ -1,7 +1,9 @@
 # inti 0.5.9
 
 - Fix for dev dplyr (Thanks @hadley)
-- `tarpuy_plotdesign()` autoconvert as factor for plot design
+- `tarpuy_plotdesign()` 
+  - Autoconvert as factor for plot design
+  - Default names in plot to "row" and "columns"
 - Yupana - fieldbook module:
   - [ ] Create a filter function and module: `yupana_filter()` 
   - [ ] Preview result?